# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/profile/preferences": {
      "put": {
        "description": "Update app-specific user preferences. Returns the updated user profile.",
        "operationId": "updatePreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User preferences object (JSON)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "dietaryRestrictions": [
                    "vegetarian"
                  ],
                  "allergies": [
                    "nuts"
                  ],
                  "theme": "dark"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User preferences updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "keycloakId": {
                      "type": "string"
                    },
                    "preferences": {
                      "type": "object"
                    },
                    "settings": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update user preferences",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/users": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid input data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource already exists or state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile",
              "roles"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all users"
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile",
              "roles"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "400": {
            "description": "Bad Request - Invalid input data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile",
              "roles"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid input data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile",
              "roles"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile",
              "roles"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}/preferences": {
      "get": {
        "operationId": "getPreferences",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User preferences"
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile",
              "roles"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get user preferences",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "updatePreferences",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences updated successfully"
          }
        },
        "security": [
          {
            "keycloak-oauth2": [
              "openid",
              "profile",
              "roles"
            ]
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update user preferences",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/pantry-item/{id}": {
      "get": {
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "pantry item found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PantryItemResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource already exists or state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get specific pantry item by ID",
        "tags": [
          "pantry-item"
        ]
      },
      "patch": {
        "description": "Update quantity, unit, notes, or expiry date of a pantry item. All fields are optional.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePantryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pantry item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PantryItemResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource already exists or state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update pantry item",
        "tags": [
          "pantry-item"
        ]
      },
      "delete": {
        "description": "Delete a specific pantry item by ID",
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "pantry item deleted successfully"
          },
          "400": {
            "description": "Bad Request - Invalid input data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource already exists or state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remove item from pantry",
        "tags": [
          "pantry-item"
        ]
      }
    },
...
  },
  "info": {
    "title": "FOODMISSION Data Framework API",
...
    "version": "1.0.0",
    "contact": {
      "name": "FOODMISSION Team",
      "url": "https://github.com/reedu-reengineering-education/foodmission-data-framework",
      "email": "support@foodmission.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.foodmission.dev",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.foodmission.dev",
      "description": "Staging server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token obtained from Keycloak authentication",
        "in": "header"
      }
    },
    "schemas": {
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 400
          },
          "message": {
            "description": "Error message or array of validation errors",
            "oneOf": [
              {
                "type": "string",
                "example": "Bad Request"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "name should not be empty",
                  "email must be a valid email"
                ]
              }
            ]
          },
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "Bad Request"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp when the error occurred",
            "example": "2024-01-15T10:30:00Z",
            "format": "date-time"
          },
          "path": {
            "type": "string",
            "description": "API path where the error occurred",
            "example": "/api/v1/foods"
          },
          "correlationId": {
            "type": "string",
            "description": "Correlation ID for request tracing",
            "example": "abc123def456"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error",
          "timestamp",
          "path"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "keycloakId": {
            "type": "string",
            "description": "Keycloak user ID"
          },
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "firstName": {
            "type": "string",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "description": "User last name"
          },
          "preferences": {
            "type": "object",
            "description": "User preferences"
          },
          "settings": {
            "type": "object",
            "description": "User settings"
          },
          "username": {
            "type": "string",
            "description": "Unique username"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Date of birth (ISO date)"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "region": {
            "type": "string",
            "description": "Region / state / province"
          },
          "zip": {
            "type": "string",
            "description": "ZIP / postal code"
          },
          "language": {
            "type": "string",
            "description": "Language preference"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "UNSPECIFIED",
              "PREFER_NOT_TO_SAY"
            ]
          },
          "annualIncome": {
            "type": "number",
            "description": "Annual income (optional)"
          },
          "educationLevel": {
            "type": "string",
            "description": "Education level"
          },
          "weightKg": {
            "type": "number",
            "description": "Weight in kg"
          },
          "heightCm": {
            "type": "number",
            "description": "Height in cm"
          },
          "activityLevel": {
            "type": "string",
            "description": "Activity level (e.g., low, medium, high)"
          },
          "healthGoals": {
            "type": "object",
            "description": "Health goals (JSON)"
          },
          "nutritionTargets": {
            "type": "object",
            "description": "Nutrition targets (JSON)"
          }
        },
        "required": [
          "keycloakId",
          "email",
          "firstName",
          "lastName"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "keycloakId": {
            "type": "string",
            "description": "Keycloak user ID"
          },
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "firstName": {
            "type": "string",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "description": "User last name"
          },
          "preferences": {
            "type": "object",
            "description": "User preferences"
          },
          "settings": {
            "type": "object",
            "description": "User settings"
          },
          "username": {
            "type": "string",
            "description": "Unique username"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Date of birth (ISO date)"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "region": {
            "type": "string",
            "description": "Region / state / province"
          },
          "zip": {
            "type": "string",
            "description": "ZIP / postal code"
          },
          "language": {
            "type": "string",
            "description": "Language preference"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "UNSPECIFIED",
              "PREFER_NOT_TO_SAY"
            ]
          },
          "annualIncome": {
            "type": "number",
            "description": "Annual income (optional)"
          },
          "educationLevel": {
            "type": "string",
            "description": "Education level"
          },
          "weightKg": {
            "type": "number",
            "description": "Weight in kg"
          },
          "heightCm": {
            "type": "number",
            "description": "Height in cm"
          },
          "activityLevel": {
            "type": "string",
            "description": "Activity level (e.g., low, medium, high)"
          },
          "healthGoals": {
            "type": "object",
            "description": "Health goals (JSON)"
          },
          "nutritionTargets": {
            "type": "object",
            "description": "Nutrition targets (JSON)"
          }
        }
      },
      "UserPreferencesDto": {
        "type": "object",
        "properties": {
          "dietaryRestrictions": {
            "description": "Dietary restrictions",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergies": {
            "description": "Food allergies",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredCategories": {
            "description": "Preferred food categories",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PantryItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the pantry item",
            "example": "uuid-item-id"
          },
          "quantity": {
            "type": "number",
            "description": "The quantity of the item",
            "example": 2,
            "minimum": 1
          },
          "unit": {
            "type": "string",
            "description": "The unit of measurement",
            "example": "KG",
            "enum": [
              "PIECES",
              "G",
              "KG",
              "ML",
              "L",
              "CUPS"
            ]
          },
          "notes": {
            "type": "string",
            "description": "Additional notes for the item",
            "example": "Store in cool place",
            "maxLength": 500
          },
          "expiryDate": {
            "format": "date-time",
            "type": "string",
            "description": "When the food will expire",
            "example": "2027-02-02T00:00:00.000Z"
          },
          "pantryId": {
            "type": "string",
            "description": "The ID of the pantry",
            "example": "uuid-pantry-id"
          },
          "foodId": {
            "type": "string",
            "description": "The ID of the food item",
            "example": "uuid-food-id"
          },
          "food": {
            "description": "The food item details",
            "allOf": [
              {
                "$ref": "#/components/schemas/FoodResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "quantity",
          "unit",
          "pantryId",
          "foodId"
        ]
      },
...
      "UpdatePantryItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "description": "The quantity of the item",
            "example": 3,
            "minimum": 0.01
          },
          "unit": {
            "type": "string",
            "description": "Unit of the item",
            "example": "KG",
            "enum": [
              "PIECES",
              "G",
              "KG",
              "ML",
              "L",
              "CUPS"
            ]
          },
          "notes": {
            "type": "string",
            "description": "Additional notes for the item",
            "example": "Buy organic if available",
            "maxLength": 500
          },
          "expiryDate": {
            "format": "date-time",
            "type": "string",
            "description": "When the food will expire (ISO date string)",
            "example": "2027-03-15"
          },
          "foodId": {
            "type": "string",
            "description": "The ID of the food item",
            "example": "uuid-food-id"
          }
        }
      }
    }
  }
}